Refactor DB. Event manager a part of DB, not the other way around. We want to add lots of DB extensions like event manager.
